docs: update README formatting
Update the formatting of the README file to improve readability and clarity. Specifically, modify the indentation and add a blockquote to a section for better emphasis. This change ensures that the ESLint configuration is correctly applied based on the `type` field in the `package.json` file. If the `type` is `module`, the ESLint configuration will be different compared to when it is not. Refer to the "CommonJS 설정" section for details on the configuration when using CommonJS.